fix: address PR review comments
- Remove duplicate import in convert_units_test.py
- Add additional offset unit test cases (0°C, 100°C)
- Add clearer error message for missing exports
- Remove redundant .astype(float) in convert_units.py
- Fix lf2013 coagulation test to check finite before nan_to_num
- Rename volume fractions test to clarify auto-normalization
- Add security comment for subprocess calls in run_linters.py
- Improve error message in create_workspace.py

+++ particula/util/convert_units.py
@@ -96,4 +96,4 @@ def get_unit_conversion(
     if result_array.shape == ():
         return result_array.item()
 
-    return result_array.astype(float)
+    return result_array
